<refactor>(domain/Name.java): Car의 name을 래핑
